+++ lib/future_context.dart
@@ -1,4 +1,6 @@
 library future_context;
 
 export 'src/future_context.dart';
+export 'src/future_context_extensions.dart';
+export 'src/stream_extensions.dart';
 export 'src/timeout_cancellation_exception.dart';